update virtual table summary token holding trades
